Extract optimization core
